[Chore #154163412] fix code climate issues
