+++ .gitignore
@@ -6,3 +6,4 @@ api/__pycache__/
 api/migrations/__pycache__/
 config/__pycache__/
 api/tests/__pycache__/
+api/views/__pycache__/
