+++ api/views/auth_views.py
@@ -1,12 +1,9 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView, status
-from rest_framework import viewsets
 from rest_framework.generics import (
     UpdateAPIView,
     CreateAPIView,
-    ListAPIView,
-    ListCreateAPIView,
-    RetrieveUpdateDestroyAPIView
+    ListAPIView
 )
 from rest_framework.permissions import (
     IsAuthenticated,
@@ -15,9 +12,6 @@ from rest_framework.permissions import (
 )
 from rest_framework_jwt.authentication import JSONWebTokenAuthentication
 from api.serializers import (
-    # UserProfileSerializer,
-    ShoppingListSerializer,
-    ItemsSerializer,
     CompositeUserSerializer,
     TokenSerializer
 )
@@ -27,10 +21,7 @@ from api.utils import (
     create_user_profile,
     update_user_profile,
     user_is_permitted,
-    serialize_item,
-    items_results_set
 )
-from api.models import ShoppingList, Item
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from rest_framework_jwt.settings import api_settings
@@ -241,207 +232,3 @@ class LogoutUser(ListAPIView):
     def get(self, request, *args, **kwargs):
         logout(request)
         return Response(status=status.HTTP_200_OK)
-
-
-class ShoppingLists(viewsets.ModelViewSet):
-    """
-    Shopping Lists CRUD endpoints
-    """
-
-    queryset = ShoppingList.objects.all()
-    serializer_class = ShoppingListSerializer
-    authentication_classes = (JSONWebTokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
-
-    # TODO: pagination
-
-    def list(self, request, *args, **kwargs):
-        """
-        get all shopping lists as per user logged in
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        lists = self.queryset.filter(user=request.user)
-        results = []
-        for a_list in lists:
-            results.append({
-                'id': a_list.id,
-                'name': a_list.name,
-                'description': a_list.description
-            })
-        return Response(ShoppingListSerializer(results, many=True).data)
-
-    def create(self, request, *args, **kwargs):
-        """
-        Adds a new shopping list
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        name = request.data.get('name', '')
-        if not name:  # shopping list name is mandatory
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-
-        new_list = ShoppingList.objects.create(
-            name=name,
-            description=request.data.get('description', ''),
-            user=request.user
-        )
-        serializer = ShoppingListSerializer(
-            data={
-                'name': new_list.name,
-                'description': new_list.description
-            })
-        serializer.is_valid()
-        return Response(
-            data=serializer.data
-        )
-
-
-class ListAllItems(ListAPIView):
-    """
-    List all items that belong to a user
-    """
-    authentication_classes = (JSONWebTokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
-
-    def get(self, request, *args, **kwargs):
-        items = Item.objects.raw(
-            "select api_item.id, api_item.name, api_item.description, "
-            "api_item.bought, api_shoppinglist.name, auth_user.first_name "
-            "from api_item "
-            "join api_shoppinglist "
-            "on api_shoppinglist.id=api_item.the_list_id "
-            "join auth_user "
-            "on auth_user.id=api_shoppinglist.user_id where auth_user.id=%s",
-            [request.user.id]
-        )
-        return Response(ItemsSerializer(
-            items_results_set(items),
-            many=True).data)
-
-
-class ItemsListCreate(ListCreateAPIView):
-    """
-    View to list and create items of/for a logged in user
-    """
-
-    authentication_classes = (JSONWebTokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
-
-    def get(self, request, *args, **kwargs):
-        """
-        Return all items for the logged in user
-
-        :param request:
-        :param kwargs:
-        :return:
-        """
-        items = Item.objects.raw(
-            "select * from api_item where "
-            "api_item.the_list_id=%s",
-            [kwargs['list_id']]
-        )
-        results = []
-        for item in items:
-            results.append({
-                'id': item.id,
-                'name': item.name,
-                'description': item.description,
-                'bought': item.bought
-            })
-        return Response(ItemsSerializer(
-            items_results_set(items),
-            many=True).data)
-
-    def post(self, request, *args, **kwargs):
-        """
-        Create a new item on a given list
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        try:
-            the_list = ShoppingList.objects.get(id=kwargs['list_id'])
-            name = request.data.get('name', '')
-            if not name:  # item name is mandatory
-                return Response(status=status.HTTP_400_BAD_REQUEST)
-            item = Item.objects.create(
-                name=name,
-                description=request.data.get('description', ''),
-                the_list=the_list
-            )
-            return Response(serialize_item(item).data)
-        except ShoppingList.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-
-class ItemsDetails(RetrieveUpdateDestroyAPIView):
-    """
-    View to retrieve, update, delete an item
-    """
-
-    queryset = Item.objects.all()
-    authentication_classes = (JSONWebTokenAuthentication,)
-    permission_classes = (IsAuthenticated,)
-
-    def get(self, request, *args, **kwargs):
-        """
-        retrieve a specific item on a given list
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        item_id = kwargs['item_id']
-        try:
-            item = Item.objects.get(id=item_id)
-            return Response(serialize_item(item).data)
-        except Item.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-    def put(self, request, *args, **kwargs):
-        """
-        update a item
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        try:
-            item = Item.objects.get(id=kwargs['item_id'])
-            name = request.data.get('name', '')
-            if not name:
-                return Response(status=status.HTTP_400_BAD_REQUEST)
-            item.name = name
-            item.description = request.data.get('description', '')
-            item.bought = request.data.get('bought', 0)
-            item.save()
-            return Response(serialize_item(item).data)
-        except Item.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-    def delete(self, request, *args, **kwargs):
-        """
-        Delete an item
-
-        :param request:
-        :param args:
-        :param kwargs:
-        :return:
-        """
-        try:
-            item = Item.objects.get(id=kwargs['item_id'])
-            item.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except Item.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
